+++ WickedEngine/wiServer.h
@@ -67,7 +67,7 @@ public:
 		time.tv_sec=0;
 		time.tv_usec=1;
 
-		if ((select( max_sd + 1 , &readfds , NULL , NULL , &time) < 0) && (errno!=EINTR))
+		if ((select( (int)max_sd + 1, &readfds , NULL , NULL , &time) < 0) && (errno!=EINTR))
 		{
 			wiBackLog::post("select error");
 		}
